requirements.txt
python-telegram-bot==20.7
PyDrive2==1.19.0
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
